fix(navigation-bar): split the class components in the App.js file into different files. split the class components into different files since react only accepts one class component per file and to easily re factor and maintain our code.
