+++ netflix-clone/src/components/Applogo.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+
+
+class AppLogo extends React.Component {
+    render () {
+      return (
+      <div className="LogoBox">
+        <a href="#logo" className='logo-text'>Netflix</a>
+      </div>
+    );
+    }
+}
+
+export default AppLogo;
