+++ netflix-clone/src/components/SubNavDrop.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+class SubNavDropdown extends React.Component {
+    render () {
+      return (
+        <div className='subnav-element'>
+          <span className='notification'>
+          <img class="profile-icon" src="https://occ-0-6517-32.1.nflxso.net/dnm/api/v6/K6hjPJd6cR6FpVELC5Pd6ovHRSk/AAAABdcxMzRQIm7B1VbSEQ6Gzbh547sBcWOewiM6BYM1RpDbqtkb-Wt_r9tEQuiI216ZlFv7tGdvn5nnY2CCYIM-TtPu_uhOVK4.png?r=e6c" alt="" />
+          </span>
+        </div>
+      )
+    }
+}
+
+export default SubNavDropdown;
